test: Add tests for findPair function
An advantage of taking a test-driven development approach is that it
ensures that tests are not being forgotten about and that the function
is meeting standards prior to release. Another benifit is increased
confidence in not introducing regressions with later refactoring.

Creating a stub in the TechnicalPrompt module just to get the tests
running for now, will bounce back over to that next. From here it's
ping-pong between the function and tests.

The final fail() within the test file is to bug future-me to look at
performance of the coded solution.