feat: Use Carton and Try::Tiny
Conceiving an idea to throw an exception inside the sort code block
while checking values but then catch that outside and return the
expected null value in those cases.

It's best to avoid using eval and to instead use Try::Tiny for something
like this. Since that is a third party library though, Carton can
facilitate managing our dependencies.